migrated snapshot tests
